pyapi/requirements.txt
Flask>=2.0.3
python-dotenv>=0.19.2
jmcomic>=2.5.22
img2pdf>=0.6.0
psutil>=5.9.0